+++ integration_tests/suite/test_status.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from hamcrest import assert_that
@@ -43,6 +43,45 @@ class TestStatusAllOK(BaseIntegrationTest):
 
         def all_connections_ok():
             result = webhookd.status.get()
-            assert_that(result['bus_consumer'], has_entries({'status': 'ok'}))
+            assert_that(
+                result,
+                has_entries(
+                    bus_consumer=has_entries(status='ok'),
+                    master_tenant=has_entries(status='ok'),
+                ),
+            )
 
         until.assert_(all_connections_ok, timeout=20)
+
+
+class TestStatusNoMasterTenant(BaseException):
+    def test_before_master_tenant_initialization(self):
+        self.stop_service('webhookd')
+        self.stop_service('auth')
+        self.start_service('webhookd')
+
+        webhookd = self.make_webhookd(MASTER_TOKEN)
+
+        def master_tenant_is_not_set():
+            result = webhookd.status.get()
+            assert_that(
+                result,
+                has_entries(
+                    master_tenant=has_entries(status='fail'),
+                ),
+            )
+
+        until.assert_(master_tenant_is_not_set, timeout=10)
+
+        self.start_service('auth')
+
+        def master_tenant_ok():
+            result = webhookd.status.get()
+            assert_that(
+                result,
+                has_entries(
+                    master_tenant=has_entries(status='ok'),
+                ),
+            )
+
+        until.assert_(master_tenant_ok, timeout=20)