+++ wazo_webhookd/plugins/status/plugin.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2021 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from .http import StatusResource
@@ -8,5 +8,8 @@ class Plugin:
     def load(self, dependencies):
         api = dependencies['api']
         bus_consumer = dependencies['bus_consumer']
+        config = dependencies['config']
 
-        api.add_resource(StatusResource, '/status', resource_class_args=[bus_consumer])
+        api.add_resource(
+            StatusResource, '/status', resource_class_args=[bus_consumer, config]
+        )