+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 21.13
 
 * The `/config` endpoint is now restricted to the master tenant
+* The `master_tenant` property has been added to the `/status` endpoint
 
 ## 21.02